Convert Audit test to spec.
